+++ src/detector.cpp
@@ -1,4 +1,4 @@
-#include <detector.h>
+#include <active_perception/detector.h>
 #include <chrono>
 
 
@@ -12,9 +12,9 @@ PinholeCamera::PinholeCamera(double fx, double fy, double cx, double cy, double 
 {
 }
 
-Eigen::Vector3f PinholeCamera::project(const Eigen::Vector3f& point_3d) const
+Eigen::Vector3d PinholeCamera::project(const Eigen::Vector3d& point_3d) const
 {
-  Eigen::Vector3f point;
+  Eigen::Vector3d point;
   point[0] = point_3d[0] * this->fx / point_3d[2] + this->cx;
   point[1] = point_3d[1] * this->fy / point_3d[2] + this->cy;
   // point[2] = point_3d[2] * 1000.0f;
@@ -22,9 +22,9 @@ Eigen::Vector3f PinholeCamera::project(const Eigen::Vector3f& point_3d) const
   return point;
 }
 
-Eigen::Vector3f PinholeCamera::backProject(const Eigen::Vector2f& point, float depth_value) const
+Eigen::Vector3d PinholeCamera::backProject(const Eigen::Vector2d& point, float depth_value) const
 {
-  Eigen::Vector3f point_3d;
+  Eigen::Vector3d point_3d;
   point_3d[2] = depth_value ;
   // point_3d[2] = depth_value / 1000.0f;
   point_3d[0] = (point[0] - this->cx) * point_3d[2] / this->fx ;
@@ -32,9 +32,9 @@ Eigen::Vector3f PinholeCamera::backProject(const Eigen::Vector2f& point, float d
   return point_3d;
 }
 
-Eigen::Matrix3f PinholeCamera::matrix() const
+Eigen::Matrix3d PinholeCamera::matrix() const
 {
-  Eigen::Matrix3f cam_matrix;
+  Eigen::Matrix3d cam_matrix;
   cam_matrix.setZero();
   cam_matrix(0, 0) = this->fx;
   cam_matrix(1, 1) = this->fy;
